pdis final delivery; kremlin stuff

+++ betweenness/betweenness/Betweenness.cpp
@@ -6,10 +6,10 @@ Betweenness::Betweenness(WeightedDirectedGraph &graph)
 	this->graph = &graph;
 }
 
-double * Betweenness::Calculate(int startVertex, int endVertex)
+int * Betweenness::Calculate(int startVertex, int endVertex)
 {
 	int vertices = graph->GetVertices();
-	double *betweenness = new double[vertices];
+	int *betweenness = new int[vertices];
 
 	for (size_t i = 0; i < vertices; i++)
 	{
@@ -18,10 +18,10 @@ double * Betweenness::Calculate(int startVertex, int endVertex)
 
 	priority_queue<KeyValuePair, vector<KeyValuePair>, greater<KeyValuePair>> Q;
 	stack<int> S;
-	double *dist = new double[vertices];
+	int *dist = new int[vertices];
 	list<int> *Pred;
-	double *sp = new double[vertices];
-	double *delta = new double[vertices];
+	int *sp = new int[vertices];
+	int *delta = new int[vertices];
 	bool *isInStack = new bool[vertices];
 
 	for (int s = startVertex; s < endVertex; s++)
@@ -35,7 +35,7 @@ double * Betweenness::Calculate(int startVertex, int endVertex)
 				Pred = new list<int>[vertices];
 				for (int i = 0; i < vertices; i++)
 				{
-					dist[i] = DBL_MAX;
+					dist[i] = INT_MAX;
 					sp[i] = 0;
 					isInStack[i] = false;
 				}
@@ -61,7 +61,7 @@ double * Betweenness::Calculate(int startVertex, int endVertex)
 					for (list<Edge*>::iterator it = adjEdges.begin(); it != adjEdges.end(); it++)
 					{
 						Edge *adjEdge = *it;
-						double edgeDistance = adjEdge->GetWeight();
+						int edgeDistance = adjEdge->GetWeight();
 						int w = adjEdge->GetOutput();
 
 						//path discovery, shorter path found?
@@ -100,9 +100,9 @@ double * Betweenness::Calculate(int startVertex, int endVertex)
 				for (list<int>::iterator it = Pred[w].begin(); it != Pred[w].end(); it++)
 				{
 					int v = *it;
-					//double before = delta[v];
+					//int before = delta[v];
 					delta[v] = delta[v] + (sp[v] / sp[w] * (1 + delta[w]));
-					//double after = delta[v];
+					//int after = delta[v];
 
 					//if (after - before > 1)
 					//{
@@ -142,12 +142,13 @@ double * Betweenness::Calculate(int startVertex, int endVertex)
 }
 
 
-double * Betweenness::CalculateOpenMP(int startVertex, int endVertex, int threads)
+int * Betweenness::CalculateOpenMP(int startVertex, int endVertex, int threads)
 {
 	//Each thread is using serial version
 	int vertices = graph->GetVertices();
-	double *betweenness = new double[vertices];
+	int *betweenness = new int[vertices];
 	omp_lock_t *lockArray = new omp_lock_t[vertices];
+
 	for (size_t i = 0; i < vertices; i++)
 	{
 		betweenness[i] = 0;
@@ -171,10 +172,11 @@ double * Betweenness::CalculateOpenMP(int startVertex, int endVertex, int thread
 		{
 			end = vertices;
 		}
+
 		//cout << "Thread: " << tid << ", start: " << start << ", end: " << end << endl;
 
 		//for low number of vertices if start is higher than max vertex id then this method will not modify betweenness
-		double *subBetweenness = Calculate(start, end);
+		int *subBetweenness = Calculate(start, end);
 		
 		for (int i = 0; i < graph->GetVertices(); i++) {
 			omp_set_lock(&lockArray[i]);
@@ -183,8 +185,8 @@ double * Betweenness::CalculateOpenMP(int startVertex, int endVertex, int thread
 		}
 
 		delete[] subBetweenness;
+
 	}
-	
 	for (size_t i = 0; i < vertices; i++)
 	{
 		omp_destroy_lock(&lockArray[i]);
